refactor AdditionalFeature to use useState
when a feature is bought, instead of having a state prop bought and
changing buy when the buyFeature action is dispatch. Create a bought
state in the feature using the state hook and toggle on and off when a
feature is added or removed, this will minimize the number of
re-renders.